-
git-svn-id: https://svn.tribler.org/abc/branches/mainbranch@23092 001aeff7-3401-0410-a489-f7902fc005dd

+++ Tribler/Core/dispersy/conversion.py
@@ -703,7 +703,7 @@ class BinaryConversion(Conversion):
         if not ceil(num_slices * bits_per_slice / 8.0) == len(data) - offset:
             raise DropPacket("Invalid number of bytes available")
         bloom_filter = BloomFilter(data, num_slices, bits_per_slice, offset=offset, prefix=prefix)
-        offset += ceil(num_slices * bits_per_slice / 8.0)
+        offset += int(ceil(num_slices * bits_per_slice / 8.0))
         
         return offset, placeholder.meta.payload.implement(destination_address, source_lan_address, source_wan_address, advice, identifier, time_low, time_high, bloom_filter)
